Move time functions into lib/time
