+++ components/nostr-auth.js
@@ -8,7 +8,7 @@ import { useRouter } from 'next/router'
 import AccordianItem from './accordian-item'
 import BackIcon from '@/svgs/arrow-left-line.svg'
 import styles from './lightning-auth.module.css'
-import { callWithTimeout } from '@/lib/nostr'
+import { callWithTimeout } from '@/lib/time'
 
 function ExtensionError ({ message, details }) {
   return (
