+++ lib/nostr.js
@@ -1,6 +1,7 @@
 import { bech32 } from 'bech32'
 import { nip19 } from 'nostr-tools'
 import WebSocket from 'isomorphic-ws'
+import { callWithTimeout, withTimeout } from '@/lib/time'
 
 export const NOSTR_PUBKEY_HEX = /^[0-9a-fA-F]{64}$/
 export const NOSTR_PUBKEY_BECH32 = /^npub1[02-9ac-hj-np-z]+$/
@@ -15,23 +16,6 @@ export const DEFAULT_CROSSPOSTING_RELAYS = [
   'wss://nostr.mutinywallet.com/',
   'wss://relay.mutinywallet.com/'
 ]
-
-function timeoutPromise (timeout) {
-  return new Promise((resolve, reject) => {
-    // if no timeout is specified, never settle
-    if (!timeout) return
-
-    setTimeout(() => reject(new Error('timeout')), timeout)
-  })
-}
-
-async function withTimeout (promise, timeout) {
-  return await Promise.race([promise, timeoutPromise(timeout)])
-}
-
-async function callWithTimeout (fn, timeout) {
-  return await Promise.race([fn(), timeoutPromise(timeout)])
-}
 
 export class Relay {
   constructor (relayUrl) {
