+++ lib/time.js
@@ -127,3 +127,20 @@ function tzOffset (tz) {
   const targetOffsetHours = (date.getTime() - targetDate.getTime()) / 1000 / 60 / 60
   return targetOffsetHours
 }
+
+function timeoutPromise (timeout) {
+  return new Promise((resolve, reject) => {
+    // if no timeout is specified, never settle
+    if (!timeout) return
+
+    setTimeout(() => reject(new Error('timeout')), timeout)
+  })
+}
+
+export async function withTimeout (promise, timeout) {
+  return await Promise.race([promise, timeoutPromise(timeout)])
+}
+
+export async function callWithTimeout (fn, timeout) {
+  return await Promise.race([fn(), timeoutPromise(timeout)])
+}
